refactor: refactoring types and schemas
