+++ src/schema/YouTubeFeedSchema.ts
@@ -30,6 +30,3 @@ export const youTubeFeedSchema = z.object({
     }),
   }),
 });
-
-// 推論される型
-export type YouTubeFeed = z.infer<typeof youTubeFeedSchema>;
