+++ src/types/YouTubeFeed.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { youTubeFeedSchema } from '../schema/YouTubeFeedSchema';
+
+export type YouTubeFeed = z.infer<typeof youTubeFeedSchema>;
